Change GoBackArrow to a component for common use
