+++ src/components/CategoryButtons/CategoryButtons.stories.js
@@ -1,6 +1,5 @@
 import React from 'react'
 import CategoryButtons from './CategoryButtons'
-import backIcon from '../../icons/back.svg'
 
 export default {
   title: 'Components/Button',
@@ -9,8 +8,4 @@ export default {
 
 export const allCategories = () => {
   return <CategoryButtons />
-}
-
-export const goBack = () => {
-  return <img src={backIcon} alt="arrow left" />
 }
