+++ src/components/GoBackArrow/GoBackArrow.stories.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import backIcon from '../../icons/back.svg'
+import GoBackArrow from './GoBackArrow'
+
+export default {
+  title: 'Components/Button',
+  component: GoBackArrow,
+}
+
+export const goBack = () => {
+  return <img src={backIcon} alt="arrow left" />
+}
